gsd-power-manager: Connect to light sensor async
Claiming the light sensor can take a noticeable amount of time when the
sensor is being polled constantly (e.g. via monitor-sensor).

Due to the function being invoked in backlight_enable this can lead to a
noticeable delay when unblanking via key press. Fix this by claiming the
sensor async.

Fixes: https://gitlab.gnome.org/World/Phosh/phosh/-/issues/915