d/g-s-d-common.lintian-overrides: Silence appstream-metadata-missing-modalias-provide
The purpose of <provides><modalias> is to get hardware support packages
to be installed on systems with specific categories of hardware.
gnome-settings-daemon is automatically installed on systems that have
the GNOME desktop environment, but doesn't make much sense to install
on systems that don't have GNOME, so it would be inappropriate to ask
for it to be installed on laptops that run some other desktop environment
and happen to have rfkill devices.